requirements.txt
fastapi
uvicorn
requests
sqlalchemy
moviepy
pillow
starlette
flask
